Coverage on Travis CI (#89)
* Coverage on Travis CI

Show coverage on Travis CI. Herewith we can improve coverage, it will prevent
this kind of bug witch library doesn't work specific version of Python.
https://github.com/gengo/gengo-python/pull/87

The coverage library doesn't support Python 3.2, however I don't think it is
big problem. Because theoretically the coverage is alomost same with other
Python's coverage. We can know the coverage to see an other result.
I think better than notthing.

I moved from requiremnts.txt to setup.py for instaalling dependencies. Because
we cannnot write version specified things in requirements.txt, but we can do it
with setup.py. We cannot install coverage with Python 3.2. So if we use
requirements.txt, CI must be failed.
We were using requirements.txt for developing and testing. We can do same thing
with:

::

   pip3 install -e .[test]

* Drop explicit support for Python 3.2

gengo-python is depends on requests. However no longer it is not supporting
Python 3.2 now.