Force absolute import
Python 3 doesn't support relative import. This is one of the reasons,
gengo-python doesn't work with Python 3. The problem was solved with this pull
request. https://github.com/gengo/gengo-python/pull/82

I forgot to add this change in my above pull request. With that line, we can
force to use absolute import.